bugfix:complie warnning

+++ frameworks/core/components_ng/pattern/tabs/tab_bar_pattern.cpp
@@ -1689,7 +1689,7 @@ void TabBarPattern::AdjustFocusPosition()
     CHECK_NULL_VOID(host);
     auto layoutProperty = host->GetLayoutProperty<TabBarLayoutProperty>();
     CHECK_NULL_VOID(layoutProperty);
-    if (focusIndicator_ < 0 || focusIndicator_ + 1 >= tabItemOffsets_.size() ||
+    if (focusIndicator_ < 0 || static_cast<uint32_t>(focusIndicator_ + 1) >= tabItemOffsets_.size() ||
         layoutProperty->GetTabBarModeValue(TabBarMode::FIXED) != TabBarMode::SCROLLABLE) {
         return;
     }
